Added functionality to UpdateLoanActivity
-Can add or deduct loans now and any changes will be reflected to the database.
-Recycler View for loan history is functional now too so the loan history can be viewed.

+++ app/src/main/java/com/example/loantracker/data/LoanHistory.java
@@ -12,7 +12,7 @@ public class LoanHistory {
     private int loan_id;
     private double amount;
 
-    public LoanHistory(int loan_id, int amount) {
+    public LoanHistory(int loan_id, double amount) {
         this.loan_id = loan_id;
         this.amount = amount;
     }
